게시글, 회원

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -1,35 +1,27 @@
 package com.example.demo.controller;
 
-import java.io.IOException;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.example.demo.service.ArticleService;
-import com.example.demo.service.BoardService;
 import com.example.demo.util.Ut;
 import com.example.demo.vo.Article;
-import com.example.demo.vo.Board;
 import com.example.demo.vo.ResultData;
 import com.example.demo.vo.Rq;
 
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 
 @Controller
 public class UsrArticleController {
-	
+
 	@Autowired
 	private Rq rq;
-	
-	@Autowired
-	private BoardService boardService;
-	
+
 	@Autowired
 	private ArticleService articleService;
 
@@ -38,123 +30,88 @@ public class UsrArticleController {
 	}
 
 	// 액션 메서드
-	@RequestMapping("/usr/article/list")
-	public String showList(HttpServletRequest req, Model model, @RequestParam(defaultValue = "1") int boardId,
-			@RequestParam(defaultValue = "1") int page,
-			@RequestParam(defaultValue = "title,body") String searchKeywordTypeCode,
-			@RequestParam(defaultValue = "") String searchKeyword) {
+	@RequestMapping("/usr/article/write")
+	public String showJoin(HttpServletRequest req) {
 
-		Rq rq = (Rq) req.getAttribute("rq");
-		Board board = boardService.getBoardById(boardId);
-		
-		int articlesCount = articleService.getArticlesCount(boardId, searchKeywordTypeCode, searchKeyword);
-		
-		if (board == null) {
-			return rq.historyBackOnView("없는 게시판이야");
-		}
-
-		int itemsInAPage = 15;
-		
-		int pagesCount = (int) Math.ceil(articlesCount / (double) itemsInAPage);
-
-		List<Article> articles = articleService.getForPrintArticles(boardId, itemsInAPage, page, searchKeywordTypeCode, searchKeyword);
-
-		model.addAttribute("board", board);
-		model.addAttribute("boardId", boardId);
-		model.addAttribute("page", page);
-		model.addAttribute("pagesCount", pagesCount);
-		model.addAttribute("searchKeywordTypeCode", searchKeywordTypeCode);
-		model.addAttribute("searchKeyword", searchKeyword);
-		model.addAttribute("articlesCount", articlesCount);
-		model.addAttribute("articles", articles);
-
-		return "usr/article/list";
+		return "usr/article/write";
 	}
-	
 	
 	@RequestMapping("/usr/article/detail")
 	public String showDetail(HttpServletRequest req, Model model, int id) {
 		Rq rq = (Rq) req.getAttribute("rq");
-		
-		Article article = articleService.getForPrintArticle(id);
-		
-//		if (!rq.isLogined()) {
-//			return Ut.jsHistoryBack("F-A", "로그아웃 상태입니다");
-//		}
-//
-//		int hit = article.getHit();
-//		int good = article.getGood();
-//		int loginedId = rq.getLoginedMemberId();
-//
-//		articleService.hitArticle(id, hit);
-//		articleService.goodArticle(id, good, loginedId);
-		
+
+		Article article = articleService.getForPrintArticle(rq.getLoginedMemberId(), id);
+
 		model.addAttribute("article", article);
 
 		return "usr/article/detail";
 	}
-	
+
+	@RequestMapping("/usr/article/list")
+	public String showList(Model model) {
+		List<Article> articles = articleService.getArticles();
+
+		model.addAttribute("articles", articles);
+
+		return "usr/article/list";
+	}
+
+	@RequestMapping("/usr/article/doWrite")
+	@ResponseBody
+	public ResultData<Article> doWrite(HttpServletRequest req, String title, String body) {
+
+		Rq rq = (Rq) req.getAttribute("rq");
+
+		if (Ut.isNullOrEmpty(title)) {
+			return ResultData.from("F-1", "제목을 입력해주세요");
+		}
+		if (Ut.isNullOrEmpty(body)) {
+			return ResultData.from("F-2", "내용을 입력해주세요");
+		}
+
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(rq.getLoginedMemberId(), title, body);
+
+		int id = (int) writeArticleRd.getData1();
+
+		Article article = articleService.getArticle(id);
+
+		return ResultData.newData(writeArticleRd, "article", article);
+	}
+
 	@RequestMapping("/usr/article/modify")
 	public String showModify(HttpServletRequest req, Model model, int id) {
 		Rq rq = (Rq) req.getAttribute("rq");
-		
-		Article article = articleService.getForPrintArticle( id);
-		
+
+		Article article = articleService.getForPrintArticle(rq.getLoginedMemberId(), id);
+
 		if (article == null) {
 			return Ut.jsHistoryBack("F-1", Ut.f("%d번 글은 존재하지 않습니다", id));
 		}
+
 		model.addAttribute("article", article);
 
 		return "usr/article/modify";
 	}
 	
-	@RequestMapping("/usr/article/write")
-	public String showWrite(HttpServletRequest req, Model model) {
-
-		return "usr/article/write";
-	}
-	
-	@RequestMapping("/usr/article/doWrite")
-	@ResponseBody
-	public String doWrite(HttpServletRequest req, Model model, String title, String body) {
-		Rq rq = (Rq) req.getAttribute("rq");
-
-		if (Ut.isNullOrEmpty(title)) {
-			return Ut.jsHistoryBack("F-1", "제목을 입력해주세요");
-		}
-		if (Ut.isNullOrEmpty(body)) {
-			return Ut.jsHistoryBack("F-1", "내용을 입력해주세요");
-		}
-		
-		ResultData<Integer> writeArticleRd = articleService.writeArticle(title, body);
-		int id = (int) writeArticleRd.getData1();
-
-		Article article = articleService.getArticle(id);
-
-		return Ut.jsReplace(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), "../article/detail?id=" + id);
-	}
-
 	// 로그인 체크 -> 유무 체크 -> 권한 체크 -> 수정
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public String doModify(HttpServletRequest req, Model model, int id, String title, String body) {
+	public ResultData<Integer> doModify(HttpServletRequest req, int id, String title, String body) {
 		Rq rq = (Rq) req.getAttribute("rq");
 
 		Article article = articleService.getArticle(id);
 
-//		if (article == null) {
-//			return Ut.jsHistoryBack("F-1", Ut.f("%d번 글은 존재하지 않습니다", id));
-//		}
+		if (article == null) {
+			return ResultData.from("F-1", Ut.f("%d번 글은 존재하지 않습니다", id), "id", id);
+		}
 
-//		ResultData loginedMemberCanModifyRd = articleService.userCanModify(rq.getLoginedMemberId(), article);
-//
-//		if (loginedMemberCanModifyRd.isSuccess()) {
-//			articleService.modifyArticle(id, title, body);
-//		}
+		ResultData loginedMemberCanModifyRd = articleService.userCanModify(rq.getLoginedMemberId(), article);
 
-		articleService.modifyArticle(id, title, body);
-		
-		return Ut.jsHistoryBack("S-1", Ut.f("%d번 글이 수정되었습니다.", id));
+		if (loginedMemberCanModifyRd.isSuccess()) {
+			articleService.modifyArticle(id, title, body);
+		}
+
+		return ResultData.from(loginedMemberCanModifyRd.getResultCode(), loginedMemberCanModifyRd.getMsg(), "id", id);
 	}
 
 	// 로그인 체크 -> 유무 체크 -> 권한 체크 -> 삭제
@@ -164,21 +121,19 @@ public class UsrArticleController {
 		Rq rq = (Rq) req.getAttribute("rq");
 
 		Article article = articleService.getArticle(id);
-//
-//		if (article == null) {
-//			return Ut.jsHistoryBack("F-1", Ut.f("%d번 글은 존재하지 않습니다", id));
-//		}
-//
-//		ResultData loginedMemberCanDeleteRd = articleService.userCanDelete(rq.getLoginedMemberId(), article);
-//		
-//		if (loginedMemberCanDeleteRd.isSuccess()) {
-//			articleService.deleteArticle(id);
-//		}
-		
-		articleService.deleteArticle(id);
-		
-		//return Ut.jsHistoryBack("S-1", Ut.f("%d번 글이 삭제되었습니다.", id));
-		return Ut.jsReplace("S-1",Ut.f("%d번 글이 삭제되었습니다.", id), "../home/main");
+
+		if (article == null) {
+			return Ut.jsHistoryBack("F-1", Ut.f("%d번 글은 존재하지 않습니다", id));
+		}
+
+		ResultData loginedMemberCanDeleteRd = articleService.userCanDelete(rq.getLoginedMemberId(), article);
+
+		if (loginedMemberCanDeleteRd.isSuccess()) {
+			articleService.deleteArticle(id);
+		}
+
+		return Ut.jsReplace(loginedMemberCanDeleteRd.getResultCode(), loginedMemberCanDeleteRd.getMsg(),
+				"../article/list");
 	}
 
 }
